+++ common/api/src/main/java/com/dfsek/terra/api/util/generic/Monoid.java
@@ -1,5 +1,8 @@
 package com.dfsek.terra.api.util.generic;
 
+import com.dfsek.terra.api.util.generic.kinds.K;
+
+
 public interface Monoid<T, M extends K<M, T>> extends Semigroup<T, M>{
     <T1, M1 extends K<M1, T1>> Monoid<T1, M1> identity();
