+++ common/api/src/main/java/com/dfsek/terra/api/util/generic/Semigroup.java
@@ -1,5 +1,8 @@
 package com.dfsek.terra.api.util.generic;
 
+import com.dfsek.terra.api.util.generic.kinds.K;
+
+
 public interface Semigroup<T, S extends K<S, T>> {
     Semigroup<T, S> multiply(S t);
 }
