+++ common/api/src/main/java/com/dfsek/terra/api/util/generic/kinds/K2.java
@@ -1,0 +1,7 @@
+package com.dfsek.terra.api.util.generic.kinds;
+
+/**
+ * Kind of the type T<A, B>
+ */
+public interface K2<T, A, B> {
+}
